Implement the logic of the process_orders_endpoint. (#24)
* feat: add process order_logic

* feat: block the execution of the process order since we dont have accses to the price of the assets

* feat: add the reply endpoint logic

* feat: remove debug comment

+++ src/action/query/get_order.rs
@@ -1,7 +1,7 @@
 use super::*;
 use crate::{bindings::query::ElysQuery, msg::query_resp::GetOrderResp};
 
-pub fn get_order(deps: Deps<ElysQuery>, order_id: u128) -> Result<GetOrderResp, ContractError> {
+pub fn get_order(deps: Deps<ElysQuery>, order_id: u64) -> Result<GetOrderResp, ContractError> {
     let orders_list: Vec<Order> = ORDER.load(deps.storage)?;
     let have_order = orders_list.iter().find(|order| order.order_id == order_id);
